requirements.txt
aws-cdk-lib==2.144.0
constructs>=10.0.0,<11.0.0
setuptools==70.1.1
# Optional
# pre-commit==3.7.1
